Refactor code to improve user profile fetching
This commit revamps how user profiles and stats are fetched. Now, specific controllers and database helper functions are created to handle fetching user profiles and stats, where previously everything was done within the route handler. This improves code modularity and keeps route handlers clean.

In addition, a validateUser middleware was introduced. This middleware validates user id, checks if the user exists in the database, and returns a response if the user cannot be found. This reduces the number of database calls within route handlers and avoids code duplication.

Rate limiters in AuthRouter has been updated to differentiate production and non-production environments.

The JSONStringify helper function has also been introduced to handle BigInt values when stringifying objects for JSON response, ensuring compatibility across different environments.

This commit greatly improves code organization and readability, and it boosts the performance of the application.