Remove unused model files and refactor existing model classes
This commit removes unnecessary files related to unused models including `Tag`, `TabInfo` and refactors existing model classes, `User`, `RoadmapView`, `Roadmap`, and `Comment`. The refactoring includes updating the class attributes to encapsulation methodology (private attributes and getter/setter methods), and updating class constructors to take in full object parameters rather than individual properties. This change was necessary to improve the project's maintainability and code readability. It reduces the risk of bugs in the long term and enhances conformity with the principles of object-oriented programming.