Enhance data model clarity, apply defaults for building objects
In response to a need for improved clarity, labels were updated for better distinction between complete object interfaces and construction interfaces across all model files. Also, default values were implemented in constructors to allow the creation of incomplete model objects. This change aims to enhance flexibility during object creation and as well to improve readability and maintainability of the codebase.