Fix default isPublic value in roadmapController
The previous code was allowing user input to affect the value of isPublic in the Roadmap controller. This commit fixes this by ensuring that isPublic is always true, as currently, there shouldn't be any non-public roadmaps. The feature for users to modify the isPublic status will be introduced in the future.