Add version field to Roadmap model
A 'version' field has been added to the 'Roadmap' model and setup.sql script. This field can be used to track the version number of each roadmap. A RegEx check has been added to make sure that the version follows the correct format (e.g. '1.0.0'). By default, the version is set to be '0.0.0'. This change will be beneficial for future roadmap iterations and version controls to ensure backwards compatibility.