Add 'RoadmapProgress' model and database schema
Added the 'RoadmapProgress' model to the types/model directory and updated the database schema in the setup.sql with the new roadmapProgress table. These additions enable the tracking of a user's progress in roadmaps.

The 'RoadmapProgress' model includes methods for constructing and modifying a RoadmapProgress object as well as a method for confirming an object is of type 'IRoadmapProgress'. The database schema specifies the structure of the roadmapProgress table, including the id, roadmapId, userId, data, createdAt and updatedAt fields, their types, and the table indexes.