Moved 'like' and 'dislike' routes to update, add progress tracking to roadmap
Removed routes for 'like' and 'dislike' functionality from the RoadmapsRouter and added them to RoadmapsUpdate router. This is in order to streamline the API to the roadmap update process. In addition, new routes and methods were added to handle roadmap progress tracking, enabling users to easily track their progress through a roadmap. These changes improve the user experience by integrating all functionality related to a roadmap (owner only updates, likes/dislikes, progress tracking) into one router. For the progress tracking, new methods were introduced for getting, inserting, and updating a user's roadmap progress in the database, and a new 'progress' route was added to the roadmaps path constants, allowing for progress data to be sent with a POST request.