Handle null values in ExploreDB queries
Made changes in ExploreDB.ts to handle null values in SQL queries, specifically in the counting of 'likeCount' and 'isLiked'. This is important to ensure data integrity and avoid runtime errors when these fields are not available.

Also, limited the 'isLiked' subquery to one result to prevent unnecessary data. As a debugging measure, console.log has been added to log the 'result'. This will help monitor the output of the queries, but should be removed in production.