Modify query parameter handling in ExploreDB
Fixed condition that pushes userid and topic into params in ExploreDB.ts. The changes ensure that the userid only gets added when it is not a countQuery and there is a valid userid. Additionally, the logic for handling an array of topics has been updated to spread the array values instead of pushing the entire array.