Update session validation and paths for like/dislike
Changes include switching to 'RoadmapsUpdate.all' in RoadmapsUpdate.ts to apply 'validateSession' for all types of HTTP requests. This change is deemed necessary for better session validation.

Also in Paths.ts, the Like and Dislike paths were streamlined from '/:roadmapId([0-9]+)/like' and '/:roadmapId([0-9]+)/dislike' to '/like' and '/dislike' respectively to simplify their structure. Please note that in this case, the roadmapId will have to be passed in the request body or as a query parameter.