Add profanity filter utilities
Implemented a profanity filter under utility directory. This includes a ProfanityFilterUtil.ts file consisting of functions to check strings and objects for profane words. It is accompanied by a WordList.ts file that contains a list of bad word patterns. These additions are aimed at preventing misuse of our platform by limiting language that could be offensive or inappropriate.