Prevent publishing drafts when roadmap is unlisted for profanity.
This commit adds a check in the roadmapController to disallow publishing a roadmap when the roadmap is not public. This also adds a new response sent by calling the new responseCantPublishDraft function, which sends an error message indicating that a draft cannot be published if it's unlisted because of inappropriate content.