Update minimum Node version and update README
This commit updates the minimum required NodeJS version from 16 to 18 in the package.json file. This change was necessary due to the new features/dependencies introduced in the project that require NodeJS 18 or higher. In addition, the README file got updated to reflect this requirement. It now indicates that NodeJS 18 (LTS) or higher is necessary. MariaDB version also updated to 10.11 (LTS) for clarity.