Add JavaScript and TypeScript migrations support
Added JavaScript and TypeScript migrations support to sql files migration system by creating new interface 'MigrationJsFile.ts' and modifying 'DatabaseDriver.ts' to handle '.js' and '.ts' migrations.

Also, 'tsconfig.json' was modified to understand .js files and 'eslintrc.json' was modified to give warnings instead of errors on ES syntax.

DatabaseDriver.ts was also modified to provide more public functionality and to run migrations and exit the process in case of failure. To facilitate these changes, a dummy migration file '1-jsExample.ts' was added.

Purpose of this change is to enhance the existing migration system and ensuring that it can handle more scenarios.